requirements.txt

requests>=2.32
SQLAlchemy>=2.0
psycopg2-binary
tabulate
python-dotenv>=1.0
